@@ -1,10 +1,11 @@
+#pragma once
 #if !defined(_LOCAL_ALLOC_H_)
 #define _LOCAL_ALLOC_H_
 
+#include <vector>
+
 #include "Common.h"
 #include "GlobalAddress.h"
-
-#include <vector>
 
 // for fine-grained shared memory alloc
 // not thread safe
@@ -12,7 +13,6 @@
 // TODO: slab-based alloctor
 class LocalAllocator
 {
-
 public:
     LocalAllocator()
     {
@@ -22,7 +22,6 @@ public:
 
     GlobalAddress malloc(size_t size, bool &need_chunck, bool align = false)
     {
-
         if (align)
         {
         }
@@ -39,7 +38,7 @@ public:
             cur.offset += size;
         }
 
-        // assert(res.addr + size <= 40 * define::GB);
+        // assert(res.addr + size <= 40_GB);
 
         return res;
     }
@@ -50,7 +49,7 @@ public:
         head = cur = addr;
     }
 
-    void free(const GlobalAddress &addr)
+    void free(const GlobalAddress &)
     {
         // TODO
     }
