@@ -1,22 +1,26 @@
 /**
  * @file rdma
  */
+#pragma once
 #ifndef _RDMA_H__
 #define _RDMA_H__
 
 #define forceinline inline __attribute__((always_inline))
 
 #include <assert.h>
+#include <glog/logging.h>
 #include <infiniband/verbs.h>
 #include <stdint.h>
 #include <stdlib.h>
 #include <unistd.h>
 
 #include <cstring>
+#include <functional>
 #include <list>
+#include <optional>
 #include <string>
 
-#include "Debug.h"
+#include "Common.h"
 
 #define MAX_POST_LIST 32
 #define DCT_ACCESS_KEY 3185
@@ -53,10 +57,38 @@ struct RdmaContext
     union ibv_gid gid;
     ibv_mw_type mw_type;
 
+    ibv_exp_dm *dm{nullptr};
+
+    ibv_exp_res_domain *res_doms[kMaxAppThread] = {};
+
     // default to use type 1, because it is always supported
     // will query and change to TYPE_2 if supported.
     RdmaContext() : ctx(NULL), pd(NULL), mw_type(IBV_MW_TYPE_1)
     {
+    }
+    RdmaContext &operator=(const RdmaContext &) = delete;
+    RdmaContext(const RdmaContext &) = delete;
+    RdmaContext &operator=(RdmaContext &&rhs)
+    {
+        devIndex = std::move(rhs.devIndex);
+        port = std::move(rhs.port);
+        gidIndex = std::move(rhs.gidIndex);
+        ctx = std::move(rhs.ctx);
+        rhs.ctx = nullptr;
+        pd = std::move(rhs.pd);
+        rhs.pd = nullptr;
+        lid = std::move(rhs.lid);
+        gid = std::move(rhs.gid);
+        mw_type = std::move(rhs.mw_type);
+        dm = std::move(rhs.dm);
+        rhs.dm = nullptr;
+        memcpy(res_doms, rhs.res_doms, sizeof(res_doms));
+        memset(rhs.res_doms, 0, sizeof(rhs.res_doms));
+        return *this;
+    }
+    RdmaContext(RdmaContext &&rhs)
+    {
+        *this = std::move(rhs);
     }
 };
 
@@ -69,31 +101,56 @@ struct Region
 };
 
 //// Resource.cpp
-bool createContext(RdmaContext *context,
-                   uint8_t port = 1,
-                   int gidIndex = 1,
-                   uint8_t devIndex = 0);
-bool destoryContext(RdmaContext *context);
+bool createContext(
+    RdmaContext *context,
+    uint8_t port = 1,
+    int gidIndex = 1,
+    uint8_t devIndex = 0,
+    std::optional<ibv_exp_thread_model> thread_model = std::nullopt,
+    std::optional<ibv_exp_msg_model> msg_model = std::nullopt);
+bool destroyContext(RdmaContext *context);
 
-ibv_mr *createMemoryRegion(uint64_t mm, uint64_t mmSize, RdmaContext *ctx);
+ibv_mr *createMemoryRegion(uint64_t mm,
+                           uint64_t mmSize,
+                           const RdmaContext *ctx);
+bool destroyMemoryRegion(ibv_mr *mr);
+bool reregisterMemoryRegionAccess(ibv_mr *mr, int access, RdmaContext *ctx);
+bool reregisterMemoryRegionTranslate(ibv_mr *mr,
+                                     void *mm,
+                                     size_t mm_size,
+                                     RdmaContext *ctx);
+
 ibv_mr *createMemoryRegionOnChip(uint64_t mm,
                                  uint64_t mmSize,
                                  RdmaContext *ctx);
+bool destroyMemoryRegionOnChip(struct ibv_mr *mr, struct ibv_exp_dm *dm);
 
 bool createQueuePair(ibv_qp **qp,
                      ibv_qp_type mode,
                      ibv_cq *cq,
                      RdmaContext *context,
-                     uint32_t qpsMaxDepth = 128,
-                     uint32_t maxInlineData = 0);
+                     uint32_t qpsMaxDepth,
+                     uint32_t maxInlineData,
+                     ibv_exp_res_domain *res_dom);
 
 bool createQueuePair(ibv_qp **qp,
                      ibv_qp_type mode,
                      ibv_cq *send_cq,
                      ibv_cq *recv_cq,
                      RdmaContext *context,
-                     uint32_t qpsMaxDepth = 128,
-                     uint32_t maxInlineData = 0);
+                     uint32_t qpsMaxDepth,
+                     uint32_t maxInlineData,
+                     ibv_exp_res_domain *res_dom);
+bool createQueuePair(ibv_qp **qp,
+                     ibv_qp_type mode,
+                     ibv_cq *send_cq,
+                     ibv_cq *recv_cq,
+                     RdmaContext *context,
+                     size_t max_send_wr,
+                     size_t max_recv_wr,
+                     uint32_t maxInlineData,
+                     ibv_exp_res_domain *res_dom);
+bool destroyQueuePair(ibv_qp *qp);
 
 bool createDCTarget(ibv_exp_dct **dct,
                     ibv_cq *cq,
@@ -106,6 +163,15 @@ void fillAhAttr(ibv_ah_attr *attr,
                 RdmaContext *context);
 
 //// StateTrans.cpp
+/**
+ * used in error handling
+ */
+bool modifyErrQPtoNormal(struct ibv_qp *qp,
+                         uint32_t remoteQPN,
+                         uint16_t remoteLid,
+                         const uint8_t *remoteGid,
+                         RdmaContext *context);
+bool modifyQPtoReset(struct ibv_qp *qp);
 bool modifyQPtoInit(struct ibv_qp *qp, RdmaContext *context);
 bool modifyQPtoRTR(struct ibv_qp *qp,
                    uint32_t remoteQPN,
@@ -120,7 +186,25 @@ bool modifyDCtoRTS(struct ibv_qp *qp,
                    uint8_t *remoteGid,
                    RdmaContext *context);
 
+ibv_cq *createCompleteQueue(RdmaContext *context,
+                            int cqe,
+                            ibv_exp_res_domain *);
+bool destroyCompleteQueue(ibv_cq *cq);
+
 //// Operation.cpp
+using WcHandler = std::function<void(ibv_wc *)>;
+using WcErrHandler = WcHandler;
+static WcErrHandler empty_wc_err_handler = [](ibv_wc *) {};
+static WcHandler empty_wc_handler = [](ibv_wc *) {};
+static WcHandler log_wc_handler = [](ibv_wc *wc) {
+    if (unlikely(wc->status != IBV_WC_SUCCESS))
+    {
+        LOG(ERROR) << "[wc] Failed status " << ibv_wc_status_str(wc->status)
+                   << " (" << wc->status << ") for wr_id " << WRID(wc->wr_id)
+                   << " at QP: " << wc->qp_num
+                   << ". vendor err: " << wc->vendor_err;
+    }
+};
 /**
  * @brief block and poll the CQ until the specified number
  * @param cq the CQ to polled with
@@ -128,7 +212,11 @@ bool modifyDCtoRTS(struct ibv_qp *qp,
  * @param wc the polling results
  * @return the number of wc actually polled.
  */
-int pollWithCQ(ibv_cq *cq, int pollNumber, struct ibv_wc *wc);
+int pollWithCQ(ibv_cq *cq,
+               int pollNumber,
+               struct ibv_wc *wc,
+               const WcErrHandler &hander = empty_wc_err_handler,
+               const WcHandler &handler = empty_wc_handler);
 /**
  * @brief non-blocking and try to poll the CQ and return immediately if get
  * nothing.
@@ -156,7 +244,10 @@ bool rdmaSend(ibv_qp *qp,
               uint32_t lkey,
               ibv_ah *ah,
               uint32_t remoteQPN,
-              bool isSignaled = false);
+              bool isSignaled = false,
+              bool isInlined = false,
+              uint64_t wr_id = 0,
+              std::optional<uint32_t> imm = std::nullopt);
 
 /**
  * @brief rdma send for RC and UC
@@ -167,6 +258,9 @@ bool rdmaSend(ibv_qp *qp,
               uint64_t source,
               uint64_t size,
               uint32_t lkey,
+              bool signal,
+              bool inlined,
+              uint64_t wr_id,
               int32_t imm = -1);
 /**
  * @brief rdma receive
@@ -347,7 +441,9 @@ bool rdmaBatchWrite(ibv_qp *qp,
                     int32_t imm = -1);
 
 //// Utility.cpp
-void rdmaQueryQueuePair(ibv_qp *qp);
+ibv_qp_state rdmaQueryQueuePair(ibv_qp *qp);
+void rdmaReportQueuePair2(ibv_qp *qp);
+void rdmaReportQueuePair(ibv_qp *qp);
 void checkDctSupported(struct ibv_context *ctx);
 /**
  * print out the information of devices
@@ -378,9 +474,106 @@ bool rdmaWriteCas(ibv_qp *qp,
                   bool isSignaled,
                   uint64_t wrID = 0);
 
-constexpr int IBV_ACCESS_CUSTOM_REMOTE_ALL =
+/**
+ * @brief fill SGE and WR and links the SGE to the WR
+ */
+static inline void fillSgeWr(
+    ibv_sge &sg, ibv_send_wr &wr, uint64_t source, uint64_t size, uint32_t lkey)
+{
+    memset(&wr, 0, sizeof(wr));
+
+    // optimized for zero-size op, e.g., zero-size SEND
+    if (size == 0)
+    {
+        wr.num_sge = 0;
+    }
+    else
+    {
+        memset(&sg, 0, sizeof(sg));
+        sg.addr = (uintptr_t) source;
+        sg.length = size;
+        sg.lkey = lkey;
+
+        wr.wr_id = 0;
+        wr.sg_list = &sg;
+        wr.num_sge = 1;
+    }
+}
+
+static inline void fillSgeWr(
+    ibv_sge &sg, ibv_recv_wr &wr, uint64_t source, uint64_t size, uint32_t lkey)
+{
+    memset(&wr, 0, sizeof(wr));
+
+    if (size == 0)
+    {
+        wr.num_sge = 0;
+    }
+    else
+    {
+        memset(&sg, 0, sizeof(sg));
+        sg.addr = (uintptr_t) source;
+        sg.length = size;
+        sg.lkey = lkey;
+
+        wr.wr_id = 0;
+        wr.sg_list = &sg;
+        wr.num_sge = 1;
+    }
+}
+
+static inline void fillSgeWr(ibv_sge &sg,
+                             ibv_exp_send_wr &wr,
+                             uint64_t source,
+                             uint64_t size,
+                             uint32_t lkey)
+{
+    memset(&wr, 0, sizeof(wr));
+    if (size == 0)
+    {
+        wr.num_sge = 0;
+    }
+    else
+    {
+        memset(&sg, 0, sizeof(sg));
+        sg.addr = (uintptr_t) source;
+        sg.length = size;
+        sg.lkey = lkey;
+
+        wr.wr_id = 0;
+        wr.sg_list = &sg;
+        wr.num_sge = 1;
+    }
+}
+inline void fill_bind_mw_wr(ibv_send_wr &wr,
+                            ibv_mw *mw,
+                            ibv_mr *mr,
+                            uint64_t addr,
+                            size_t length,
+                            int access_flag)
+{
+    wr.sg_list = nullptr;
+    wr.num_sge = 0;
+    wr.opcode = IBV_WR_BIND_MW;
+    wr.imm_data = 0;
+    wr.wr_id = 0;
+    wr.send_flags = 0;
+    wr.bind_mw.mw = mw;
+    wr.bind_mw.rkey = mw->rkey;
+    wr.bind_mw.bind_info.mr = mr;
+    wr.bind_mw.bind_info.addr = addr;
+    wr.bind_mw.bind_info.length = length;
+    wr.bind_mw.bind_info.mw_access_flags = access_flag;
+}
+
+constexpr int IBV_ACCESS_CUSTOM_REMOTE_RW =
     IBV_ACCESS_LOCAL_WRITE | IBV_ACCESS_REMOTE_WRITE | IBV_ACCESS_REMOTE_READ |
-    IBV_ACCESS_REMOTE_ATOMIC;
+    IBV_ACCESS_REMOTE_ATOMIC | IBV_ACCESS_MW_BIND;
+constexpr int IBV_ACCESS_CUSTOM_REMOTE_RO =
+    IBV_ACCESS_LOCAL_WRITE | IBV_ACCESS_REMOTE_READ | IBV_ACCESS_REMOTE_ATOMIC |
+    IBV_ACCESS_MW_BIND;
+constexpr int IBV_ACCESS_CUSTOM_REMOTE_NORW =
+    IBV_ACCESS_LOCAL_WRITE | IBV_ACCESS_REMOTE_ATOMIC | IBV_ACCESS_MW_BIND;
 
 /**
  * bind a memory window to a part of memory region.
@@ -396,6 +589,7 @@ uint32_t rdmaAsyncBindMemoryWindow(
     struct ibv_mr *mr,
     uint64_t mm,
     uint64_t mmSize,
+    bool signal,
     uint64_t wrID = 0,
-    unsigned int mw_access_flag = IBV_ACCESS_CUSTOM_REMOTE_ALL);
+    unsigned int mw_access_flag = IBV_ACCESS_CUSTOM_REMOTE_RW);
 #endif
