@@ -1,14 +1,14 @@
+#pragma once
 #if !defined(_SIMPLE_HT_H_)
 #define _SIMPLE_HT_H_
+
+#include <map>
 
 #include "Common.h"
 #include "GlobalAddress.h"
 
-#include <map>
-
 class SimpleHT
 {
-
     const static uint64_t kBuckSize = 640000;
 
 public:
@@ -53,7 +53,7 @@ private:
         Key k;
         GlobalAddress v;
     };
-    Item *table;
+    [[maybe_unused]] Item *table;
     std::map<Key, GlobalAddress> t;
 };
 
