@@ -1,10 +1,20 @@
 #include "Keeper.h"
 
+#include <glog/logging.h>
+
+#include <chrono>
 #include <fstream>
 #include <iostream>
 #include <random>
+#include <sstream>
+
+#include "util/Util.h"
+
+using namespace std::chrono_literals;
 
 char *getIP();
+
+thread_local PerThreadMemcachedSt Keeper::thread_memc;
 
 std::string trim(const std::string &s)
 {
@@ -19,7 +29,7 @@ std::string trim(const std::string &s)
 
 const char *Keeper::SERVER_NUM_KEY = "serverNum";
 
-Keeper::Keeper(uint32_t maxServer) : maxServer(maxServer), memc(NULL)
+Keeper::Keeper(uint32_t maxServer) : maxServer(maxServer), global_memc(NULL)
 {
 }
 
@@ -42,7 +52,7 @@ bool Keeper::connectMemcached()
 
     if (!conf)
     {
-        error("can't open memchaced.conf at ../memcached.conf");
+        LOG(ERROR) << "can't open memchaced.conf at ../memcached.conf";
         return false;
     }
 
@@ -50,28 +60,31 @@ bool Keeper::connectMemcached()
     std::getline(conf, addr);
     std::getline(conf, port);
 
-    memc = memcached_create(NULL);
+    global_memc = memcached_create(NULL);
     servers = memcached_server_list_append(
         servers, trim(addr).c_str(), std::stoi(trim(port)), &rc);
-    rc = memcached_server_push(memc, servers);
+    rc = memcached_server_push(global_memc, servers);
+
+    free(servers);
 
     if (rc != MEMCACHED_SUCCESS)
     {
-        error("Can't add server:%s", memcached_strerror(memc, rc));
+        LOG(ERROR) << "Can't add server:"
+                   << memcached_strerror(global_memc, rc);
         return false;
     }
 
-    memcached_behavior_set(memc, MEMCACHED_BEHAVIOR_BINARY_PROTOCOL, 1);
+    memcached_behavior_set(global_memc, MEMCACHED_BEHAVIOR_BINARY_PROTOCOL, 1);
     return true;
 }
 
 bool Keeper::disconnectMemcached()
 {
-    if (memc)
+    if (global_memc)
     {
-        memcached_quit(memc);
-        memcached_free(memc);
-        memc = NULL;
+        memcached_quit(global_memc);
+        memcached_free(global_memc);
+        global_memc = NULL;
     }
     return true;
 }
@@ -84,17 +97,23 @@ void Keeper::serverEnter()
     while (true)
     {
         rc = memcached_increment(
-            memc, SERVER_NUM_KEY, strlen(SERVER_NUM_KEY), 1, &serverNum);
+            global_memc, SERVER_NUM_KEY, strlen(SERVER_NUM_KEY), 1, &serverNum);
         if (rc == MEMCACHED_SUCCESS)
         {
             myNodeID = serverNum - 1;
+            auto overwrite_node_id = util::node_id_in_rank_file();
+            if (overwrite_node_id.has_value())
+            {
+                myNodeID = overwrite_node_id.value();
+            }
 
-            dinfo("This server get NodeId %d [%s]\n", myNodeID, getIP());
+            DLOG(INFO) << "This server get NodeId " << myNodeID << "["
+                       << getIP() << "]";
             return;
         }
-        error("Server %d Counld't incr value and get ID: %s, retry...",
-              myNodeID,
-              memcached_strerror(memc, rc));
+        LOG(ERROR) << "Server " << myNodeID
+                   << " Counld't incr value and get ID: "
+                   << memcached_strerror(global_memc, rc) << ". retry. ";
         sleep(1);
     }
 }
@@ -108,13 +127,16 @@ void Keeper::serverConnect()
 
     while (curServer < maxServer)
     {
-        char *serverNumStr = memcached_get(
-            memc, SERVER_NUM_KEY, strlen(SERVER_NUM_KEY), &l, &flags, &rc);
+        char *serverNumStr = memcached_get(global_memc,
+                                           SERVER_NUM_KEY,
+                                           strlen(SERVER_NUM_KEY),
+                                           &l,
+                                           &flags,
+                                           &rc);
         if (rc != MEMCACHED_SUCCESS)
         {
-            error("Server %d Counld't get serverNum: %s, retry\n",
-                  myNodeID,
-                  memcached_strerror(memc, rc));
+            LOG(ERROR) << "Server " << myNodeID << " Counld't get serverNum:"
+                       << memcached_strerror(global_memc, rc) << ". retry. ";
             sleep(1);
             continue;
         }
@@ -127,68 +149,9 @@ void Keeper::serverConnect()
             if (k != myNodeID)
             {
                 connectNode(k);
-                dinfo("Connected to server %zu", k);
+                DLOG(INFO) << "Connected to server " << k;
             }
         }
         curServer = serverNum;
-    }
-}
-
-void Keeper::memSet(const char *key,
-                    uint32_t klen,
-                    const char *val,
-                    uint32_t vlen)
-{
-    memcached_return rc;
-    while (true)
-    {
-        rc =
-            memcached_set(memc, key, klen, val, vlen, (time_t) 0, (uint32_t) 0);
-        if (rc == MEMCACHED_SUCCESS)
-        {
-            break;
-        }
-        sleep(400);
-        std::this_thread::yield();
-    }
-}
-
-char *Keeper::memGet(const char *key, uint32_t klen, size_t *v_size)
-{
-    size_t l;
-    char *res;
-    uint32_t flags;
-    memcached_return rc;
-
-    while (true)
-    {
-        res = memcached_get(memc, key, klen, &l, &flags, &rc);
-        if (rc == MEMCACHED_SUCCESS)
-        {
-            break;
-        }
-        usleep(400 * myNodeID);
-    }
-
-    if (v_size != nullptr)
-    {
-        *v_size = l;
-    }
-
-    return res;
-}
-
-uint64_t Keeper::memFetchAndAdd(const char *key, uint32_t klen)
-{
-    uint64_t res;
-    while (true)
-    {
-        memcached_return rc = memcached_increment(memc, key, klen, 1, &res);
-        if (rc == MEMCACHED_SUCCESS)
-        {
-            return res;
-        }
-        usleep(10000);
-        std::this_thread::yield();
     }
 }
