@@ -1,11 +1,11 @@
+#include "Common.h"
 #include "DSM.h"
 #include "Timer.h"
 
 // Two nodes
 // one node issues cas operations
 
-constexpr uint16_t kTargetNodeId = 1;
-constexpr uint32_t kMachineNr = 2;
+constexpr uint16_t kTargetNodeId = 0;
 
 int main()
 {
@@ -14,7 +14,7 @@ int main()
     rdmaQueryDevice();
 
     DSMConfig config;
-    config.machineNR = kMachineNr;
+    config.machineNR = ::config::kMachineNr;
 
     auto dsm = DSM::getInstance(config);
 
@@ -34,7 +34,7 @@ int main()
     gaddr.nodeID = kTargetNodeId;
     gaddr.offset = 1024;
 
-    auto *buffer = dsm->get_rdma_buffer();
+    auto *buffer = dsm->get_rdma_buffer().buffer;
 
     {
         uint64_t cur_val = 0;
@@ -108,74 +108,14 @@ int main()
 
         printf("\n-------- cas and read succ ----------\n");
 
-        RdmaOpRegion cas_ror;
-        cas_ror.dest = gaddr;
-        cas_ror.is_on_chip = true;
-        cas_ror.source = (uint64_t) buffer;
-        cas_ror.size = 8;
-
-        RdmaOpRegion read_ror;
-        read_ror.dest = GADD(gaddr, 1024);
-        read_ror.is_on_chip = false;
-        read_ror.source = (uint64_t) buffer + 8;
-        read_ror.size = 8;
-
-        timer.begin();
-        for (size_t i = 0; i < loop; ++i)
-        {
-            auto cas_ror_input = cas_ror;
-            auto read_ror_input = read_ror;
-            dsm->cas_read_sync(
-                cas_ror_input, read_ror_input, cur_val, cur_val + 1);
-            cur_val++;
-        }
-        timer.end_print(loop);
-
-        printf("\n-------- write 2 succ ----------\n");
-        timer.begin();
-        for (size_t i = 0; i < loop; ++i)
-        {
-            RdmaOpRegion rs[2];
-            rs[0] = cas_ror;
-            rs[1] = read_ror;
-            dsm->write_batch_sync(rs, 2);
-            // assert(res);
-            // cur_val++;
-        }
-        timer.end_print(loop);
-
-        printf("\n-------- write faa succ ----------\n");
-        timer.begin();
-        for (size_t i = 0; i < loop; ++i)
-        {
-            auto faa_ror_input = cas_ror;
-            auto write_ror_input = read_ror;
-            dsm->write_faa_sync(write_ror_input, faa_ror_input, 1);
-            // assert(res);
-            // cur_val++;
-        }
-        timer.end_print(loop);
-
-        printf("\n-------- write cas succ ----------\n");
-        timer.begin();
-        for (size_t i = 0; i < loop; ++i)
-        {
-            auto cas_ror_input = cas_ror;
-            auto write_ror_input = read_ror;
-            dsm->write_cas_sync(
-                write_ror_input, cas_ror_input, cur_val, cur_val + 1);
-            // assert(res);
-            // cur_val++;
-        }
-        timer.end_print(loop);
-
         // function call time
         size_t call_loop = 100;
         printf("\n -------- cas function call ----------\n");
         timer.begin();
         for (size_t i = 0; i < call_loop; ++i)
         {
-            dsm->cas(gaddr, 1, cur_val + 1, (uint64_t *) buffer);
+            dsm->cas(
+                gaddr, 1, cur_val + 1, (uint64_t *) buffer, true, 0, nullptr);
         }
         timer.end_print(call_loop);
 
